+++ db/migrate/20171121205738_add_reviewer_uploaded_to_supporting_documents.rb
@@ -1,0 +1,6 @@
+class AddReviewerUploadedToSupportingDocuments < ActiveRecord::Migration[5.1]
+  def change
+    add_column :supporting_documents, :reviewer_uploaded, :boolean, null: false, default: false
+    add_index :supporting_documents, :reviewer_uploaded
+  end
+end