+++ CHANGELOG.md
@@ -88,6 +88,7 @@
   - Users can now directly upload profile pictures
 - **Reviewer Changes**
   - Simplified voting on agreements as a reviewer
+  - Reviewers can now upload supporting documents to a data request
 - **Gem Changes**
   - Updated to ruby 2.4.2
   - Updated to rails 5.1.4